requirements.txt
fastapi
uvicorn
mysql-connector-python
pytest
httpx
databases
aiomysql